Partial classes for specific target frameworks
