+++ src/Utilities.Test/Attributes/JsonTestMethodAttribute.cs
@@ -7,17 +7,8 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Matr.Utilities.Test.Attributes
 {
-    [AttributeUsage(System.AttributeTargets.Method)]
-    public class JsonTestMethodAttribute : TestMethodAttribute
+    public partial class JsonTestMethodAttribute
     {
-        private readonly string filepath;
-        private readonly Type dataType;
-        public JsonTestMethodAttribute(string filepath, Type dataType)
-        {
-            this.filepath = filepath;
-            this.dataType = dataType;
-        }
-
         public override TestResult[] Execute(ITestMethod testMethod)
         {
             using (StreamReader r = new StreamReader(filepath))
@@ -25,7 +16,11 @@ namespace Matr.Utilities.Test.Attributes
                 string json = r.ReadToEnd();
 
                 var dataArray = JsonSerializer.Deserialize(json, dataType.MakeArrayType());
+#if NET48
+                return (dataArray as IEnumerable).Cast<object>()
+#else
                 return (dataArray as IEnumerable)!.Cast<object>()
+#endif
                     .Select(x => testMethod.Invoke(new object[] { x }))
                     .ToArray();
             }
