+++ src/Utilities.Test/Attributes/JsonTestMethodAttribute_462.cs
@@ -8,17 +8,8 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Matr.Utilities.Test.Attributes
 {
-    [AttributeUsage(System.AttributeTargets.Method)]
-    public class JsonTestMethodAttribute : TestMethodAttribute
+    public partial class JsonTestMethodAttribute
     {
-        private readonly string filepath;
-        private readonly Type dataType;
-        public JsonTestMethodAttribute(string filepath, Type dataType)
-        {
-            this.filepath = filepath;
-            this.dataType = dataType;
-        }
-
         public override TestResult[] Execute(ITestMethod testMethod)
         {
             using (StreamReader r = new StreamReader(filepath))
