+++ src/Utilities.Test/Utilities.Test.csproj
@@ -21,7 +21,7 @@
     <PropertyGroup>
       <MajorVersion>0</MajorVersion>
       <MinorVersion>0</MinorVersion>
-      <PatchVersion>$([System.DateTime]::Now.ToString("yyMMdd"))</PatchVersion>
+      <PatchVersion>1</PatchVersion>
       <BuildVersion Condition="'$(DesignTimeBuild)' != 'true' OR '$(BuildingProject)' == 'true'">$([System.DateTime]::Now.ToString("HHmmss"))</BuildVersion>
       <PackageVersion>$(MajorVersion).$(MinorVersion).$(PatchVersion).$(BuildVersion)</PackageVersion>
     </PropertyGroup>
@@ -33,18 +33,26 @@
 
   <ItemGroup>
     <PackageReference Include="autofac" Version="6.2.0" />
-    <Compile Include="TestBase.cs" />
     <PackageReference Include="MSTest.TestFramework" Version="2.2.5" />
+    <PackageReference Condition=" '$(TargetFramework)' == 'net48' " Include="system.text.json" Version="5.0.2" />
+    <PackageReference Condition=" '$(TargetFramework)' == 'net462' " Include="Newtonsoft.Json" Version="13.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="TestBase.cs" />
+    <Compile Include="GenericFactory_base.cs" />
+    <Compile Include="Attributes\JsonTestMethodAttribute_base.cs" />
+
+    <Compile Condition=" '$(TargetFramework)' != 'net462' " Include="Attributes\JsonTestMethodAttribute.cs" />
+    <Compile Condition=" '$(TargetFramework)' == 'net462' " Include="Attributes\JsonTestMethodAttribute_462.cs" />
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'net6.0' OR '$(TargetFramework)' == 'net5.0' OR '$(TargetFramework)' == 'netcoreapp3.1'  ">
     <Compile Include="GenericFactory.cs" />
-    <Compile Include="Attributes\JsonTestMethodAttribute.cs" />
   </ItemGroup>
+
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'net462' OR '$(TargetFramework)' == 'net48' OR '$(TargetFramework)' == 'netstandard2.1' OR '$(TargetFramework)' == 'netstandard2.0'">
     <Compile Include="GenericFactory_462.cs" />
-    <Compile Include="Attributes\JsonTestMethodAttribute_462.cs" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
   </ItemGroup>
 </Project>
